* FULL REPORT - WORKS NOW FOR JUST THE 2 TABLES, BUT PLOTS, MAP, AND BIBLIOG FAIL
By removing the bibliography from the .Rmd header I got the Word doc to at least render, albeit with missing images and bibliog.  There is an issue with the paths / folders - it cannot find the plots files or bibliography file when run in the app. If I knit from within RStudio via the knit button, it does pull in the placeholder images and render the word doc. In the app though it renders the word doc with 2 tables but not the map not the plot(s).