PREPARE TO USE packrat TO ENABLE DEPLOYMENT-see notes
our RStudio Connect server uses packrat to manage dependencies. EJAM depends on packages not on CRAN and not on public/open github repositories, so it needs renv or packrat to be able to install those from a local repo with tarball build
or else one could deal with the token or whatever to access the private github repos or else make them public.