FIXES TO STATE AND US PERCENTILES LOOKUP
Fixed issue where wrong percentile was reported for a score that is also the minimum score in that state or US and multiple percentiles in lookup table show that same score since many places are tied at the minimum score.
Note this is NOT fixed to address where multiple percentiles are tied at a value that is NOT the minimum value... might want to fix that some day.