** Tried to fix state_from_latlon() to handle more than just 1 latlon pair - But when should it return NULL?
If it handles only 1 at a time, it can return NULL for the point if lat or lon is bad.
But if it now/again handles a vector of latlon, what if some are OK and some are not? Do we still want to return NULL? or a vector with some NA values?